+++ scripts/module.mjs
@@ -2,6 +2,7 @@ import * as settings from "./config/settings.config.mjs";
 import * as coverHandler from "./handlers/cover.hooks.mjs";
 import * as queries from "./services/queries.service.mjs";
 import * as debug from "./services/cover.debug.mjs";
+import * as hover from "./services/hover.service.mjs";
 
 
 // === Init Phase ===
@@ -10,14 +11,15 @@ Hooks.once("init", () => {
   queries.initQueries();
   coverHandler.ignoreCoverProperties();
 });
-Hooks.once("ready", () => {
+Hooks.once("canvasReady", () => {
   debug.clearCoverDebug();
 });
-
+Hooks.on('getSceneControlButtons', settings.getSceneControlButtons);
 
 // === Calc Cover Hooks ===
-Hooks.on("updateCombat", coverHandler.clearCoverOnUpdateCombat); // Clear Cover
-Hooks.on("deleteCombat", coverHandler.clearCoverOnDeleteCombat); // Clear Cover
-Hooks.on("moveToken", coverHandler.clearCoverOnMovement); // Clear Cover
-Hooks.on("dnd5e.preRollAttack", coverHandler.onPreRollAttack); // Calc Cover on attack roll
-Hooks.on("dnd5e.preRollSavingThrow", coverHandler.onPreRollSavingThrow); // Calc Cover on saving throw
+Hooks.on("updateCombat", coverHandler.clearCoverOnUpdateCombat); 
+Hooks.on("deleteCombat", coverHandler.clearCoverOnDeleteCombat);
+Hooks.on("moveToken", coverHandler.clearCoverOnMovement); 
+Hooks.on("dnd5e.preRollAttack", coverHandler.onPreRollAttack);
+Hooks.on("dnd5e.preRollSavingThrow", coverHandler.onPreRollSavingThrow); 
+Hooks.on("hoverToken", hover.onHoverToken);